Fork refinery crates refinery* as trailbase-refinery*.
This is necessary due to upstream refinery packages being very stale and
being version incompatible with the already stale upstream rusqlite.
We would love to rely on upstream refinery but the project seems to be
in hibernation right now. We might have had to fork eventually anyway or
move to a different DB migration library.